zna-streamlit-backend/zna-streamlit/requirements.txt
# ── LLM ───────────────────────────────────────────────────────────────────────
google-generativeai>=0.7.0

# ── PDF ────────────────────────────────────────────────────────────────────────
reportlab>=4.2.0

# ── Retry / Resilience ─────────────────────────────────────────────────────────
tenacity>=8.3.0

# ── Streamlit ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Data ───────────────────────────────────────────────────────────────────────
pandas>=2.0.0

# ── File parsing ───────────────────────────────────────────────────────────────
pymupdf>=1.24.0
python-docx>=1.1.0

# ── Validation ─────────────────────────────────────────────────────────────────
pydantic>=2.8.0
